feat: wire up compiler
