+++ formily/next/package.json
@@ -64,7 +64,8 @@
     "@designable/core": "1.0.0-beta.45",
     "@designable/formily-setters": "1.0.0-beta.45",
     "@designable/formily-transformer": "1.0.0-beta.45",
-    "@designable/react": "1.0.0-beta.45"
+    "@designable/react": "1.0.0-beta.45",
+    "@samagra-x/formily-compiler": "0.0.1"
   },
   "publishConfig": {
     "access": "public"
