+++ formily/next/playground/compiler/fieldTypes.js
@@ -1,0 +1,7 @@
+export const INPUT_FIELD_TYPES = {
+  INPUT: 'Input',
+  SELECT: 'Select',
+  TEXTAREA: 'Input.TextArea',
+}
+
+// module.exports = INPUT_FIELD_TYPES
