+++ formily/next/playground/widgets/ActionsWidget.tsx
@@ -32,7 +32,10 @@ export const ActionsWidget = observer(() => {
           GlobalRegistry.setDesignerLanguage(e.target.value)
         }}
       />
-      <Button href="https://github.com/alibaba/designable" target="_blank">
+      <Button
+        href="https://github.com/techsavvyash/formily-designable"
+        target="_blank"
+      >
         <GithubOutlined />
         Github
       </Button>
