--- packages/compiler/test/formily/formily_input1.json
@@ -1,24 +1,0 @@
-{
-  "form": {
-    "labelCol": 6,
-    "wrapperCol": 12
-  },
-  "schema": {
-    "type": "object",
-    "properties": {
-      "7klm1s4rvbt": {
-        "type": "string",
-        "title": "Hello",
-        "x-decorator": "FormItem",
-        "x-component": "Input",
-        "x-validator": [],
-        "x-component-props": {},
-        "x-decorator-props": {},
-        "x-designable-id": "7klm1s4rvbt",
-        "x-index": 0,
-        "description": "say Hello"
-      }
-    },
-    "x-designable-id": "r9wd0dsf4rl"
-  }
-}
