--- packages/compiler/test/formily/formily_input2.json
@@ -1,48 +1,0 @@
-{
-  "form": {
-    "labelCol": 6,
-    "wrapperCol": 12
-  },
-  "schema": {
-    "type": "object",
-    "properties": {
-      "qrxhfgc3vqp": {
-        "type": "string",
-        "title": "FirstName",
-        "x-decorator": "FormItem",
-        "x-component": "Input",
-        "x-validator": [],
-        "x-component-props": {},
-        "x-decorator-props": {},
-        "x-designable-id": "qrxhfgc3vqp",
-        "x-index": 0,
-        "description": "First Name"
-      },
-      "apy72yxa6ve": {
-        "type": "string",
-        "title": "LastName",
-        "x-decorator": "FormItem",
-        "x-component": "Input",
-        "x-validator": [],
-        "x-component-props": {},
-        "x-decorator-props": {},
-        "x-designable-id": "apy72yxa6ve",
-        "x-index": 1,
-        "description": "Last Name"
-      },
-      "hdf85u4rk5m": {
-        "type": "string",
-        "title": "FathersName",
-        "x-decorator": "FormItem",
-        "x-component": "Input",
-        "x-validator": [],
-        "x-component-props": {},
-        "x-decorator-props": {},
-        "x-designable-id": "hdf85u4rk5m",
-        "x-index": 2,
-        "description": "Father's Name"
-      }
-    },
-    "x-designable-id": "66113zcg0s6"
-  }
-}
