--- packages/compiler/test/formily/formily_select1.json
@@ -1,35 +1,0 @@
-{
-  "form": {
-    "labelCol": 6,
-    "wrapperCol": 12
-  },
-  "schema": {
-    "type": "object",
-    "properties": {
-      "62q7t4xz4kk": {
-        "title": "Class",
-        "x-decorator": "FormItem",
-        "x-component": "Select",
-        "x-validator": [],
-        "x-component-props": {},
-        "x-decorator-props": {},
-        "enum": [
-          {
-            "children": [],
-            "value": "class-1-key",
-            "label": "Class 1"
-          },
-          {
-            "children": [],
-            "value": "class-2-key",
-            "label": "Class 2"
-          }
-        ],
-        "x-designable-id": "62q7t4xz4kk",
-        "x-index": 0,
-        "description": "Select Your Class"
-      }
-    },
-    "x-designable-id": "66113zcg0s6"
-  }
-}
