+++ packages/react/package.json
@@ -35,7 +35,8 @@
   "dependencies": {
     "@designable/core": "1.0.0-beta.45",
     "@designable/shared": "1.0.0-beta.45",
-    "dateformat": "^4.5.1"
+    "dateformat": "^4.5.1",
+    "@samagra-x/formily-compiler": "0.0.1"
   },
   "publishConfig": {
     "access": "public"
